Default google client ID for authorized origins
