+++ package.json
@@ -21,6 +21,7 @@
     "@babel/preset-typescript": "7.16.7",
     "@rollup/plugin-json": "4.1.0",
     "@rollup/plugin-node-resolve": "13.1.3",
+    "@rollup/plugin-replace": "4.0.0",
     "@rollup/plugin-typescript": "8.3.1",
     "@types/jest": "27.4.1",
     "@types/node": "17.0.23",
