+++ packages/app/src/utils.ts
@@ -1,6 +1,5 @@
 import { Patient, Reference, Resource } from '@medplum/fhirtypes';
-
-declare const google: unknown;
+import { createScriptTag } from '@medplum/ui';
 
 export function getPatient(resource: Resource): Patient | Reference<Patient> | undefined {
   if (resource.resourceType === 'Patient') {
@@ -18,16 +17,6 @@ export function getPatient(resource: Resource): Patient | Reference<Patient> | u
 }
 
 /**
- * Dynamically loads the Google Auth script.
- * We do not want to load the script on page load unless the user needs it.
- */
-export function initGoogleAuth(): void {
-  if (typeof google === 'undefined') {
-    createScriptTag('https://accounts.google.com/gsi/client');
-  }
-}
-
-/**
  * Dynamically loads the recaptcha script.
  * We do not want to load the script on page load unless the user needs it.
  */
@@ -35,18 +24,6 @@ export function initRecaptcha(): void {
   if (typeof grecaptcha === 'undefined') {
     createScriptTag('https://www.google.com/recaptcha/api.js?render=' + process.env.RECAPTCHA_SITE_KEY);
   }
-}
-
-/**
- * Dynamically creates a script tag for the specified JavaScript file.
- * @param src The JavaScript file URL.
- */
-function createScriptTag(src: string): void {
-  const head = document.getElementsByTagName('head')[0];
-  const script = document.createElement('script');
-  script.async = true;
-  script.src = src;
-  head.appendChild(script);
 }
 
 /**
